Create Manipulator util
